Refactor: created constants folder and moved files accordingly.
