--- lib/rapid-hotel-api/routes.ts
@@ -1,3 +1,0 @@
-export const ROUTE_BROWSING_LANDING_PAGE = "/searchtest" as const;
-
-
